Prevent/reduce multiple triggers
This change will reset our timeout timer when a timeout
event is triggered. This is to prevent/reduce multiple triggers
if the target machine has a long running command in play
and  doesn't process the temperature reset immediately.
If the timeout delay is short enough and the target machine
has still not reset the temperature, duplicate events will still
trigger every <timeout_delay> seconds.